docker.requirements.txt
pandas
openpyxl
youtube-dl
opencv-python
termcolor
numpy
pillow
wget==3.2
tensorflow==1.12.0
keras==2.2.4
torch==1.0.1.post2
torchvision==0.2.2

### GCLOUD
google-api-core==1.8.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-cloud-vision==0.36.0
google-images-download==2.5.0
